node-agent/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
psutil>=5.9.0
gputil>=1.4.0
pyyaml>=6.0
httpx>=0.26.0
